+++ src/core/controllers/index.ts
@@ -4,3 +4,4 @@ export * from "./FPSController.js";
 export * from "./OrbitController.js";
 export * from "./WASDController.js";
 export * from "./ZoomController.js";
+export * from "./DeviceOrientationController.js";